Added call script search feature

+++ src/call-scripts/call-scripts.controller.ts
@@ -7,6 +7,7 @@ import {
   ParseUUIDPipe,
   Patch,
   Post,
+  Query,
   UseGuards,
 } from '@nestjs/common';
 import { CallScriptsService } from './call-scripts.service';
@@ -19,6 +20,7 @@ import { RequirePermissions } from 'src/roles/decorators/permissions.decorator';
 import { Permission } from 'src/roles/enums/roles-permissions.enum';
 import { ActiveSubscriptionsGuard } from 'src/subscriptions/guards/active-subscriptions.guard';
 import { UpdateCallScriptDto } from './dto/update-call-script.dto';
+import { ListCallScriptDto } from './dto/list-call-scripts.dto';
 
 @Controller('call-scripts')
 export class CallScriptsController {
@@ -31,6 +33,7 @@ export class CallScriptsController {
     @Body() createCallScriptDto: CreateCallScriptDto,
     @CurrentUser() loggedInUser: User,
   ) {
+    if (!loggedInUser.organization_id) return;
     return this.callScriptsService.create(
       createCallScriptDto,
       loggedInUser.organization_id,
@@ -41,8 +44,15 @@ export class CallScriptsController {
   @UseGuards(JwtAuthGuard, PermissionsGuard, ActiveSubscriptionsGuard)
   @RequirePermissions(Permission.EDIT_PATIENTS)
   @Get()
-  findAll(@CurrentUser() loggedInUser: User) {
-    return this.callScriptsService.findAll(loggedInUser.organization_id);
+  findAll(
+    @CurrentUser() loggedInUser: User,
+    @Query() listCallScriptsDto: ListCallScriptDto,
+  ) {
+    if (!loggedInUser.organization_id) return;
+    return this.callScriptsService.findAll(
+      loggedInUser.organization_id,
+      listCallScriptsDto,
+    );
   }
 
   @UseGuards(JwtAuthGuard, PermissionsGuard, ActiveSubscriptionsGuard)
@@ -52,6 +62,7 @@ export class CallScriptsController {
     @Param('id', ParseUUIDPipe) id: string,
     @CurrentUser() loggedInUser: User,
   ) {
+    if (!loggedInUser.organization_id) return;
     return this.callScriptsService.findOne(id, loggedInUser.organization_id);
   }
 
@@ -63,6 +74,7 @@ export class CallScriptsController {
     @Body() updateCallScriptDto: UpdateCallScriptDto,
     @CurrentUser() loggedInUser: User,
   ) {
+    if (!loggedInUser.organization_id) return;
     return this.callScriptsService.update(
       id,
       updateCallScriptDto,
@@ -78,6 +90,7 @@ export class CallScriptsController {
     @Param('id', ParseUUIDPipe) id: string,
     @CurrentUser() loggedInUser: User,
   ) {
+    if (!loggedInUser.organization_id) return;
     return this.callScriptsService.remove(id, loggedInUser.organization_id);
   }
 }
